backend/requirements.txt
moviepy
gTTS
pyttsx3
google-genai